add unified simulation script
